+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.2] - 2023-09-03
+
+### Added
+- Support for Palladiom Wire-Free shades.
+
+### Fixed
+- Restored support for Wall Mounted Motion Sensor on RA3 23.4 firmware.
+
 ## [0.18.1] - 2023-02-03
 
 ### Fixed
@@ -270,7 +278,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to interact with Caséta bridge using LIP over Telnet and LEAP over SSH.
 
-[unreleased]: https://github.com/gurumitts/pylutron-caseta/compare/v0.18.1...HEAD
+[unreleased]: https://github.com/gurumitts/pylutron-caseta/compare/v0.18.2...HEAD
+[0.18.2]: https://github.com/gurumitts/pylutron-caseta/compare/v0.18.1...v0.18.2
 [0.18.1]: https://github.com/gurumitts/pylutron-caseta/compare/v0.18.0...v0.18.1
 [0.18.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.17.1...v0.18.0
 [0.17.1]: https://github.com/gurumitts/pylutron-caseta/compare/v0.17.0...v0.17.1